test: Expand testing for bump using rstest and cases
- parameterized test and documented test cases
- for typical and custom commit types
- for production and non-production version numbers
- created placeholder file to retain tmp directory in git
- removed redundant build.rs
