--- tests/tmp/test-63638549-729b-48d1-8725-270c7d8a6ff5/test.txt
@@ -1,1 +1,0 @@
-Hello, world!
